// packages/client/src/types.ts
export enum types {
  CREATE_ROOM = 'CREATE_ROOM',
  ROOM_CREATED = 'ROOM_CREATED',
  JOIN_ROOM = 'JOIN_ROOM',
  ROOM_JOINED = 'ROOM_JOINED',
  UPDATE_FIELD = 'UPDATE_FIELD',
  UPDATE_STATE = 'UPDATE_STATE',
  UPDATE_PLAYERS = 'UPDATE_PLAYERS',
  UPDATE_CHOICES = 'UPDATE_CHOICES',
  START_SESSION = 'START_SESSION',
  SESSION_STARTED = 'SESSION_STARTED',
  PLAY_CARD = 'PLAY_CARD',
  NEW_ROUND = 'NEW_ROUND',
  START_ROUND = 'START_ROUND',
}